Added basic logit generation.
