--- fsdp2/fsdp2_example.py
@@ -1,96 +1,0 @@
-import argparse
-import os
-
-import torch
-from checkpoint import Checkpointer
-import torch.distributed as dist
-# from fsdp2.model_manual import ModelArgs, Transformer
-from torch.distributed.fsdp import fully_shard, MixedPrecisionPolicy
-from utils import inspect_mixed_precision, inspect_model
-from tqdm.auto import tqdm
-
-from transformers import AutoModelForCausalLM
-
-def set_modules_to_forward_prefetch(model, num_to_forward_prefetch):
-    for i, layer in enumerate(model.layers):
-        if i >= len(model.layers) - num_to_forward_prefetch:
-            break
-        layers_to_prefetch = [
-            model.layers[i + j] for j in range(1, num_to_forward_prefetch + 1)
-        ]
-        layer.set_modules_to_forward_prefetch(layers_to_prefetch)
-
-
-def set_modules_to_backward_prefetch(model, num_to_backward_prefetch):
-    for i, layer in enumerate(model.layers):
-        if i < num_to_backward_prefetch:
-            continue
-        layers_to_prefetch = [
-            model.layers[i - j] for j in range(1, num_to_backward_prefetch + 1)
-        ]
-        layer.set_modules_to_backward_prefetch(layers_to_prefetch)
-
-
-def main(args):
-    rank = int(os.environ["LOCAL_RANK"])
-    device = torch.device(f"cuda:{rank}")
-    torch.cuda.set_device(device)
-    torch.distributed.init_process_group(backend="nccl", device_id=device)
-    torch.manual_seed(0)
-    with torch.device("meta"):
-        # model = Transformer(model_args)
-        model = AutoModelForCausalLM.from_pretrained(
-            "Qwen/Qwen2.5-0.5B-Instruct",
-            torch_dtype=torch.bfloat16,
-        )
-    fsdp_kwargs = {}
-    if args.mixed_precision:
-        fsdp_kwargs["mp_policy"] = MixedPrecisionPolicy(
-            param_dtype=torch.bfloat16,
-            reduce_dtype=torch.float32,
-        )
-    for layer in model.model.layers:
-        fully_shard(layer, **fsdp_kwargs)
-    fully_shard(model, **fsdp_kwargs)
-
-    inspect_model(model)
-
-    if args.explicit_prefetching:
-        set_modules_to_forward_prefetch(model, num_to_forward_prefetch=2)
-        set_modules_to_backward_prefetch(model, num_to_backward_prefetch=2)
-
-    checkpointer = Checkpointer("checkpoints", dcp_api=args.dcp_api)
-    if checkpointer.last_training_time is None:
-        model.to_empty(device="cuda")
-        # model.reset_parameters()
-    else:
-        checkpointer.load_model(model)
-    
-    if args.mixed_precision:
-        inspect_mixed_precision(model)
-
-    optim = torch.optim.Adam(model.parameters(), lr=1e-2)
-    if checkpointer.last_training_time is not None:
-        checkpointer.load_optim(model, optim)
-
-    for _ in tqdm(range(100)):
-        if args.explicit_prefetching:
-            model.unshard()
-        x = torch.randint(0, 151936, (4, 1024), device=device)
-        loss = 10000 - model(x).logits.sum()
-        loss.backward()
-        torch.nn.utils.clip_grad_norm_(model.parameters(), max_norm=1.0)
-        optim.step()
-        optim.zero_grad()
-
-    checkpointer.save(model, optim)
-    torch.distributed.destroy_process_group()
-
-
-if __name__ == "__main__":
-    parser = argparse.ArgumentParser(description="PyTorch FSDP2 example")
-    parser.add_argument("--explicit-prefetching", action="store_true", default=False)
-    parser.add_argument("--mixed-precision", action="store_true", default=False)
-    parser.add_argument("--dcp-api", action="store_true", default=False)
-    args = parser.parse_args()
-    main(args)
