+++ shard_weight.py
